feat: dynamic pipelines: delete pipelines

+++ ldi-orchestrator/ldio-application/src/main/java/be/vlaanderen/informatievlaanderen/ldes/ldio/services/PipelineService.java
@@ -1,11 +1,13 @@
 package be.vlaanderen.informatievlaanderen.ldes.ldio.services;
 
 import be.vlaanderen.informatievlaanderen.ldes.ldio.config.PipelineConfig;
+import be.vlaanderen.informatievlaanderen.ldes.ldio.events.PipelineDeletedEvent;
 import be.vlaanderen.informatievlaanderen.ldes.ldio.exception.PipelineAlreadyExistsException;
 import be.vlaanderen.informatievlaanderen.ldes.ldio.exception.PipelineException;
 import be.vlaanderen.informatievlaanderen.ldes.ldio.repositories.PipelineFileRepository;
 import be.vlaanderen.informatievlaanderen.ldes.ldio.repositories.PipelineRepository;
 import be.vlaanderen.informatievlaanderen.ldes.ldio.valueobjects.PipelineConfigTO;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
@@ -13,11 +15,13 @@ import java.util.List;
 
 @Service
 public class PipelineService {
+	private final ApplicationEventPublisher applicationEventPublisher;
 	private final PipelineCreatorService pipelineCreatorService;
 	private final PipelineRepository pipelineRepository;
 
-	public PipelineService(PipelineCreatorService pipelineCreatorService,
+	public PipelineService(ApplicationEventPublisher applicationEventPublisher, PipelineCreatorService pipelineCreatorService,
 	                       PipelineFileRepository pipelineRepository) {
+		this.applicationEventPublisher = applicationEventPublisher;
 		this.pipelineCreatorService = pipelineCreatorService;
 		this.pipelineRepository = pipelineRepository;
 	}
@@ -47,6 +51,12 @@ public class PipelineService {
 	}
 
 	public boolean deletePipeline(String pipeline) {
-		return pipelineRepository.delete(pipeline);
+		if (pipelineRepository.exists(pipeline)) {
+			applicationEventPublisher.publishEvent(new PipelineDeletedEvent(pipeline));
+			pipelineCreatorService.removePipeline(pipeline);
+			return pipelineRepository.delete(pipeline);
+		} else {
+			return false;
+		}
 	}
 }
